+++ Android/app/src/main/java/com/example/justdoit/miyamoto/Unit/OkHttpSample.kt
@@ -90,41 +90,4 @@ class OkHttpSample {
             }
         })
     }
-
-    fun postLoginData(context: Context, email: String, pass: String){
-        // todo ここでpostするデータ付与して
-        val formBody = FormBody.Builder()
-                .add("email", email)
-                .add("password", pass)
-                .build()
-
-        val request = Request.Builder()
-                .url("http://140.82.9.44:3000/auth/login")       // HTTPアクセス POST送信 テスト確認用ページ
-                .post(formBody)
-                .build()
-
-        val client = OkHttpClient()
-        client.newCall(request).enqueue(object : Callback {
-            override fun onFailure(call: Call, e: IOException) {
-                //TODO 失敗処理書く！
-            }
-
-            @Throws(IOException::class)
-            override fun onResponse(call: Call, response: Response) {
-                val res = response.body()?.string()
-                (context as LoginFormActivity).runOnUiThread{
-                    val json: JSONObject
-                    try {
-                        json = JSONObject(res)
-                        token = json.getString("token")
-                        Log.i("token",token)    //TODO SharedPreferenceでToken保存
-                    } catch (e: JSONException) {
-                        e.printStackTrace()
-                    }
-                }
-            }
-        })
-    }
-
-
 }
